feat: enable RESTful api
